Updated document parser to catch follow-on input
JSON input must be either a single value, single object or single array.  If input were to continue beyond the end of the first JSON construct then it would be ignored.  This condition should be an error. alerting the user to the existence of incorrect input which would be ignored.